requirements.txt
autobahn==21.2.2
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==3.0.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-appconf==1.0.4
django-behaviors==0.5.1
django-celery-email==3.0.0
django-cors-headers==3.7.0
django-countries==7.0
django-credit-cards==0.4.1
django-jazzmin==2.4.4
django-money==1.3
django-rest-auth==0.9.5
django-silk==4.1.0
django-versatileimagefield==2.0
djangorestframework==3.12.2
drf-flex-fields==0.8.8
drf-yasg==1.20.0
Faker==6.5.0
flake8==3.8.4
gprof2dot==2021.2.21
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==7.1.2
oauthlib==3.1.0
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.14
psycopg2==2.8.6
ptyprocess==0.7.0
py==1.10.0
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.8.0
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytest-mock==3.5.1
pytest-picked==0.4.6
python-dateutil==2.8.1
python-decouple==3.4
python-magic==0.4.18
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-mock==1.8.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sentry-sdk==0.20.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
Twisted==20.3.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zope.interface==5.2.0
pytest-cov==2.11.1
PyJWT==1.7.1
flake8-bugbear==21.3.1
